RDKB-11354 : Unknown Reboot Reason for GUI Reboot
Reason for change:
Setting of reboot reason was commented as part of restore configuration fix,
as committing the syscfg db caused overwriting of replaced db.
So updated confPhp script to set a parameter restore_reboot and
added check in the code to not to commit syscfg if it is a restore reboot.

Test Procedure:
Test 1 : Reboot device from GUI. Verify the last reboot reason once the device is up
Test 2 : Save and Restore of user Configuration from GUI.
Risks: None

+++ source/TR-181/middle_layer_src/cosa_deviceinfo_dml.c
@@ -917,34 +917,44 @@ DeviceInfo_SetParamStringValue
 	 else if (AnscEqualString(pString, "reboot_device", TRUE))
          {
                 CcspTraceInfo(("RDKB_REBOOT : RebootDevice triggered from GUI\n"));
-                /*char buf[8];
-		snprintf(buf,sizeof(buf),"%d",1);
-		               
-		if (syscfg_set(NULL, "X_RDKCENTRAL-COM_LastRebootReason", "gui-reboot") != 0) 
+
+		char buffer[8] = {0};
+		syscfg_get( NULL, "restore_reboot", buffer, sizeof(buffer));
+
+		if (strcmp(buffer, "true") != 0)
 		{
-		        AnscTraceWarning(("RDKB_REBOOT : RebootDevice syscfg_set failed GUI\n"));
-		}
-		else 
-		{
-			if (syscfg_commit() != 0) 
+			char buf[8];
+			snprintf(buf,sizeof(buf),"%d",1);
+
+			if (syscfg_set(NULL, "X_RDKCENTRAL-COM_LastRebootReason", "gui-reboot") != 0)
 			{
-				AnscTraceWarning(("RDKB_REBOOT : RebootDevice syscfg_commit failed GUI\n"));
+				AnscTraceWarning(("RDKB_REBOOT : RebootDevice syscfg_set failed GUI\n"));
 			}
-		}
+			else
+			{
+				if (syscfg_commit() != 0)
+				{
+					AnscTraceWarning(("RDKB_REBOOT : RebootDevice syscfg_commit failed GUI\n"));
+				}
+			}
 	        
 	        
-	        if (syscfg_set(NULL, "X_RDKCENTRAL-COM_LastRebootCounter", buf) != 0) 
-	        {
-                    AnscTraceWarning(("syscfg_set failed\n"));
-                }
-                else 
+			if (syscfg_set(NULL, "X_RDKCENTRAL-COM_LastRebootCounter", buf) != 0)
+			{
+				AnscTraceWarning(("syscfg_set failed\n"));
+			}
+			else
+			{
+				if (syscfg_commit() != 0)
+				{
+					AnscTraceWarning(("syscfg_commit failed\n"));
+				}
+			}
+	        }
+		else
 		{
-		      if (syscfg_commit() != 0)
-	              {
-                         AnscTraceWarning(("syscfg_commit failed\n"));
-                      }
-	        }*/
-	        
+			CcspTraceInfo(("RDKB_REBOOT : RebootDevice to restore configuration\n"));
+		}
 	 }
          else if(AnscEqualString(pString, "factory_reset", TRUE))
          {
